refactor: :recycle: Reworked custom summary length logic to be variable length
Instead of allowing the user to input a desired word length, the feature is now changed to allow the user to pick 1 out of 3 possible summary length outputs. This is because it's very difficult, and not guaranteed, to output an appropriate output in the desired length. Thus, providing 3 summary length options is the solution to this problem - it still provides the "customisable" aspect.